Finito Archivio , Inizio Chat - Rimozione DOPPIO CHAT MESSAGE - Aggiunta css in ogni componente della chat - aggiunt scritta nessun messaggio - aggiunta elemento chat_history vuoto. - correzzione input_text. - aggiunta icona dentro - nuovo pulsante della clear chat - nuovo bellissimo bug in cui la chatInput scende a caso. DA SISTEMARE COL PROSSIMO PUSH
