+++ 2-RTB/poc/src/app/documents/page.tsx
@@ -4,7 +4,7 @@ import ListDoc from "./components/listDoc";
 
 export default function Page(){
     return (
-        <main id='root' className="flex flex-col gap-5">
+        <main id='root' className="flex flex-col gap-5 overflow-scroll lg:overflow-hidden">
             <div className="flex flex-row justify-between ">
             <Link className="w-fit h-fit bg-blue-500 font-medium rounded-xl p-3
           hover:scale-105 hover:shadow-2xl animation duration-300 ease-out  " href="/chat">
