+++ 2-RTB/poc/src/app/globals.css
@@ -30,6 +30,9 @@
   --secondary: #c09569;
   --accent: #1a55ad;
 
+  --message-Q: #6426dc;
+  --message-A: #331997;
+
 }
 *{
   padding: 0;
@@ -93,42 +96,6 @@ a:hover {
     animation: all 0.25ms ease-in-out;
 }
 
-
-/*
-
-a {
-  font-weight: 500;
-  color: #646cff;
-}
-
-a:hover {
-  color: #535bf2;
-}
-
-
-
-h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
-}
-
-button {
-
-  color: var(--text);
-  transition: all 0.25s;
-
-}
-button:hover {
-  color: var(--accent);
-  border-color: var(--accent);
-}
-
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-
-}
-*/
 @media (prefers-color-scheme: light) {
   :root {
     --text: #0d0f12;
@@ -140,6 +107,9 @@ button:focus-visible {
     --primary: #0262f2;
     --secondary: #976b3f;
     --accent: #528de5;
+
+    --message-Q: #528de5;
+    --message-A: #2b3035;
   }
 
 }
